+++ connect4/src/model.rs
@@ -30,8 +30,8 @@ impl Mapper {
 }
 
 impl model::tensorflow::model::Mapper<GameState,Action> for Mapper {
-    fn game_state_to_input(&self, game_state: &GameState) -> Vec<Vec<Vec<f64>>> {
-        let result: Vec<Vec<Vec<f64>>> = Vec::with_capacity(6);
+    fn game_state_to_input(&self, game_state: &GameState) -> Vec<Vec<Vec<f32>>> {
+        let result: Vec<Vec<Vec<f32>>> = Vec::with_capacity(6);
 
         map_board_to_arr(game_state.p1_piece_board).iter()
             .zip(map_board_to_arr(game_state.p2_piece_board).iter())
@@ -59,17 +59,17 @@ impl model::tensorflow::model::Mapper<GameState,Action> for Mapper {
             })
     }
 
-    fn policy_metrics_to_expected_input(&self, policy_metrics: &NodeMetrics<Action>) -> Vec<f64> {
-        let total_visits = policy_metrics.visits as f64 - 1.0;
-        let result:[f64; 7] = policy_metrics.children_visits.iter().fold([0.0; 7], |mut r, p| {
-            match p.0 { Action::DropPiece(column) => r[column as usize - 1] = p.1 as f64 / total_visits };
+    fn policy_metrics_to_expected_input(&self, policy_metrics: &NodeMetrics<Action>) -> Vec<f32> {
+        let total_visits = policy_metrics.visits as f32 - 1.0;
+        let result:[f32; 7] = policy_metrics.children_visits.iter().fold([0.0; 7], |mut r, p| {
+            match p.0 { Action::DropPiece(column) => r[column as usize - 1] = p.1 as f32 / total_visits };
             r
         });
 
         result.to_vec()
     }
 
-    fn policy_to_valid_actions(&self, game_state: &GameState, policy_scores: &Vec<f64>) -> Vec<ActionWithPolicy<Action>> {
+    fn policy_to_valid_actions(&self, game_state: &GameState, policy_scores: &Vec<f32>) -> Vec<ActionWithPolicy<Action>> {
          let valid_actions_with_policies: Vec<ActionWithPolicy<Action>> = game_state.get_valid_actions().iter()
             .zip(policy_scores).enumerate()
             .filter_map(|(i, (v, p))|
